+++ src/test/suite/index.ts
@@ -6,7 +6,7 @@ export async function run(): Promise<void> {
   const mocha = new Mocha({
     ui: 'tdd',
     color: true,
-    timeout: 3000
+    timeout: 4000
   });
 
   const testsRoot = path.resolve(__dirname);